Move create form redirection logic up one level
